+++ ray-marching-lib/vec3/vec3.h
@@ -2,12 +2,12 @@
 
 typedef struct Vec3
 {
-    double x;
-    double y;
-    double z;
+    float x;
+    float y;
+    float z;
 } Vec3;
 
-static inline Vec3 vec3_create(double x, double y, double z)
+static inline Vec3 vec3_create(float x, float y, float z)
 {
     Vec3 newVector = {x, y, z};
     return newVector;
@@ -20,16 +20,16 @@ static inline void vec3_add(Vec3 *a, const Vec3 *b)
     a->z += b->z;
 }
 
-static inline void vec3_multiply_scalar(Vec3 *a, double scalar)
+static inline void vec3_multiply_scalar(Vec3 *a, float scalar)
 {
     a->x *= scalar;
     a->y *= scalar;
     a->z *= scalar;
 }
 
-static inline Vec3 vec3_multiply_scalar_copy(const Vec3 *a, double scalar)
+static inline Vec3 vec3_multiply_scalar_copy(const Vec3 *a, float scalar)
 {
     return vec3_create(a->x * scalar, a->y * scalar, a->z * scalar);
 }
 
-void vec3_step(Vec3 *a, const Vec3 *direction, double stepSize);
+void vec3_step(Vec3 *a, const Vec3 *direction, float stepSize);
